refactor: API key regeneration
